# openapi.yaml
openapi: 3.1.0
info:
  title: Doit API
  description: |
    Self-hosted task manager. Flask backend serving the compiled React frontend.

    **Authentication:** session cookie (`td_session`) set by `POST /auth/login`,
    or an API key passed as `Authorization: Bearer <TD_API_KEY>` or `X-API-Key: <TD_API_KEY>`.
    Public paths: `/login`, `/auth/login`, `/auth/logout`, `/auth/status`, PWA assets.
  version: "1.0.0"
servers:
  - url: http://localhost:5000
    description: Local dev (gunicorn / flask)
  - url: https://doit.example.com
    description: Self-hosted production
security:
  - sessionCookie: []
  - bearerAuth: []
  - apiKeyHeader: []
...
paths:
...
  /api/projects:
    get:
      tags: [projects]
      responses:
        '200':
          description: OK (inbox first, then user-visible projects)
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Project' } }
    post:
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:      { type: string }
                color:     { type: string, default: "#6366f1" }
                icon:      { type: string, default: "📁" }
                parent_id: { type: string, nullable: true, description: "Root project to nest under. One level only; the parent must not itself have a parent." }
                description: { type: string, description: free-text description }
                due_date:  { type: string, nullable: true, description: "optional deadline as YYYY-MM-DD; empty string or null leaves it unset. An unparseable value is rejected with 400." }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/projects/reorder:
    post:
      tags: [projects]
      summary: Reorder projects (inbox is ignored if included). Positions are scoped per sibling group; send one flat depth-first list.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order]
              properties:
                order: { type: array, items: { type: string }, description: project IDs in desired order }
      responses:
        '204': { description: Reordered }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/projects/{pid}:
    parameters:
      - { in: path, name: pid, required: true, schema: { type: string } }
    put:
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
    patch:
      tags: [projects]
      summary: "Update a project. A boolean `archived` is translated by the server into the `archived_at` timestamp; archiving hides the project and all of its tasks everywhere. `inbox` cannot be archived and returns 400, as does an unparseable `due_date`."
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400': { $ref: '#/components/responses/BadRequest' }
    delete:
      tags: [projects]
      summary: Delete project; its tasks fall back to inbox and its subprojects are promoted to top level. Cannot delete `inbox`.
      responses:
        '204': { description: Deleted }

  # ── Users ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: td_session
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Project:
      type: object
      properties:
        id:         { type: string }
        name:       { type: string }
        color:     { type: string }
        icon:       { type: string }
        position:   { type: integer, description: order within the sibling group }
        parent_id:  { type: string, nullable: true, description: parent project id, or null for a top-level project }
        shared:     { type: boolean }
        description: { type: string, description: free-text project description }
        due_date:   { type: string, format: date, nullable: true, description: "optional deadline (YYYY-MM-DD). Fires a once-only all-day reminder at the configured `reminder_allday_time` and surfaces on the Dashboard at-risk list. Setting it clears `deadline_notified_at`, re-arming the reminder." }
        archived_at: { type: string, format: date-time, nullable: true, readOnly: true, description: "when the project was archived, or null. Read-only; set via the boolean `archived` field on PATCH. Archiving hides the project and all of its tasks everywhere. Archived projects never notify and never appear in the at-risk list." }
        deadline_notified_at: { type: string, format: date-time, nullable: true, readOnly: true, description: "when the deadline reminder last fired, or null while still armed. Read-only; the server stamps it after dispatching and clears it whenever `due_date` is set." }
        owner_id:   { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProjectUpdate:
      type: object
      description: Only `name`, `color`, `icon`, `shared`, `parent_id`, `description`, `due_date` and the boolean `archived` are accepted; other keys are ignored.
      properties:
        name:      { type: string }
        color:     { type: string }
        icon:      { type: string }
        shared:    { type: boolean }
        parent_id: { type: string, nullable: true, description: "Set to a root project id to nest, or null to promote to top level. Rejected with 400 if it would create a second level." }
        description: { type: string, description: "free-text description; trimmed server-side" }
        due_date:  { type: string, nullable: true, description: "deadline as YYYY-MM-DD; empty string or null clears it. An unparseable value is rejected with 400. Setting this clears `deadline_notified_at`, re-arming the deadline reminder." }
        archived:  { type: boolean, description: "true archives the project (and hides all of its tasks everywhere), false restores it; the server manages the `archived_at` timestamp. `inbox` cannot be archived and is rejected with 400." }
...